soil_salinity_multi_sensor/requirements.txt
# Dependencies for Multi-Sensor Soil Salinity Inversion Project

# Deep Learning Framework
torch>=2.0.0
torchvision>=0.15.0

# Data Processing
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=1.0.0

# Geospatial Data Processing
rasterio>=1.3.0
geopandas>=0.12.0
fiona>=1.9.0
shapely>=2.0.0

# Configuration Files
pyyaml>=6.0

# Visualization
matplotlib>=3.5.0
seaborn>=0.12.0

# Progress Bar
tqdm>=4.64.0

# Optional Dependencies (for advanced features)
# xgboost>=1.7.0  # If using XGBoost as baseline model
# tensorboard>=2.10.0  # If using TensorBoard for training logs

